requirements-mwaa.txt
apache-airflow==2.10.2
apache-airflow-providers-amazon==9.0.0
dbt-core==1.8.7
dbt-duckdb==1.8.3
boto3==1.35.0
delta-spark==3.2.1
